IOS Portfolio Marc 1
Hi I'm Marc Kurtenbach, I'm a Senior at Eden Prairie High school and I have a small background in coding and computer science. Starting the class out I've browsed through many tutorials to acquire knowledge to pull together to create my own projects even when its a bit slow going. 

Project 1- Farming Clicker App
A clicker app is always a great first app to branch off in a class, especially when you can learn other tricks and code pieces that could be used in other projects. In this app I further got used to the swift Ui and coding while using multiple code tricks like creating a working timer, a click counter, multiplier count, swapping colors and pages, and the general set up for images. The clock code was a obstacle to climb because it proved to have multiple vague errors with different pieces of code until I figured out the right code to use.

Project 2-Brick Breaker Shenanigans
After the success of my Farming clicker app I wanted to test out the other way of XCode by using sprite kit. With this piece I figured out the principles of adding physics boundaries to objects, having collisions cause a response, and a simple game score system. This project was more of a test but let me get into the workings of SpriteKit more.

Other Tutorials and Workings
I am not yet a expert at all of x code so I watched some tutorials to learn some concepts I want to use on future projects. Some tutorials I worked on was A incorporation of a endless background something that could be used on a plat former project, a impulse hockey game, a joystick, and some workings of a cloud kit app.

I hope to finish up a next project soon and expand on my portfolio when I complete it.  I've been thinking of designing a new Swift app and I am in a group working on a card game project similar to Slay the Spire.